api/requirements.txt
fastapi==0.115.4
SQLAlchemy==2.0.36
uvicorn==0.32.0
python-dotenv==1.0.1
